backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-jose[cryptography]
pyotp
qrcode[pil]
deepface
tf-keras
torch
torchaudio
transformers
openai-whisper
fer
mediapipe
ultralytics
pyannote.audio
cryptography
sqlalchemy
pydantic
python-multipart
bandit
opencv-python-headless
numpy
scipy
requests
python-dotenv
httpx
pytest
pytest-asyncio
pypdf
